Refactor handleMessage : Separate message handling into different functions
