+++ commands/args-info.js
@@ -11,4 +11,4 @@ module.exports = {
     
         message.channel.send(`Command name: args-info \nArguments: ${args}`);
     },
-}
+};
