+++ commands/ping.js
@@ -5,4 +5,4 @@ module.exports = {
     execute(message, args) {
         message.channel.send('Pong.');
     },
-}
+};
